fix error

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,15 +8,15 @@
 
 int main(void)
 {
-	int num1 = 1;
-	int num2 = 2;
-	int next;
+	unsigned long int num1 = 1;
+	unsigned long int num2 = 2;
+	unsigned long int next;
 	int i;
 
-	printf("%d, ", num1);
+	printf("%lu, ", num1);
 	for (i = 1; i < 98; i++)
 	{
-		printf("%d", num2);
+		printf("%lu", num2);
 		next = num1 + num2;
 		num1 = num2;
 		num2 = next;
